unicache: Make unified cache analysis much quicker.
The original checkForConflicts function was idiotically inefficient,
looping through the instructions in a basic block to determine the
number of cache conflicts a particular instruction has, causing analysis
of long basic blocks to take much longer than necessary. Apparently,
whoever wrote that function forgot that basic blocks are contiguous.

This change updates checkForConflicts to perform in constant time,
allowing analysis of QNX to finish in 40 seconds instead of an hour.